Refactor, add tests.
