+++ process/blockProcessor.go
@@ -155,57 +155,3 @@ func (bp *blockProcessor) GetHyperBlockByNonce(nonce uint64) (*data.HyperblockAp
 	hyperblock := builder.build()
 	return &data.HyperblockApiResponse{Data: data.HyperblockApiResponsePayload{Hyperblock: hyperblock}}, nil
 }
-
-type hyperblockBuilder struct {
-	metaBlock   *data.Block
-	shardBlocks []*data.Block
-}
-
-func (builder *hyperblockBuilder) addMetaBlock(metablock *data.Block) {
-	builder.metaBlock = metablock
-}
-
-func (builder *hyperblockBuilder) addShardBlock(block *data.Block) {
-	builder.shardBlocks = append(builder.shardBlocks, block)
-}
-
-func (builder *hyperblockBuilder) build() data.Hyperblock {
-	hyperblock := data.Hyperblock{}
-	bunch := newBunchOfTxs()
-
-	bunch.collectTxs(builder.metaBlock)
-	for _, block := range builder.shardBlocks {
-		bunch.collectTxs(block)
-	}
-
-	txs := bunch.getDeduplicated()
-	hyperblock.Nonce = builder.metaBlock.Nonce
-	hyperblock.Round = builder.metaBlock.Round
-	hyperblock.Hash = builder.metaBlock.Hash
-	hyperblock.PrevBlockHash = builder.metaBlock.PrevBlockHash
-	hyperblock.Epoch = builder.metaBlock.Epoch
-	hyperblock.NumTxs = uint32(len(txs))
-	hyperblock.Transactions = txs
-
-	return hyperblock
-}
-
-type bunchOfTxs struct {
-	txs []*data.FullTransaction
-}
-
-func newBunchOfTxs() *bunchOfTxs {
-	return &bunchOfTxs{
-		txs: make([]*data.FullTransaction, 0),
-	}
-}
-
-func (bunch *bunchOfTxs) collectTxs(block *data.Block) {
-	for _, miniBlock := range block.MiniBlocks {
-		bunch.txs = append(bunch.txs, miniBlock.Transactions...)
-	}
-}
-
-func (bunch *bunchOfTxs) getDeduplicated() []*data.FullTransaction {
-	return make([]*data.FullTransaction, 0)
-}
